moved Action classes into action folder and tests also
