+++ src/models/actions/ActionBegin.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace TaskForce\models;
+namespace TaskForce\models\actions;
 
-class ActionDeny extends AbstractAction
+class ActionBegin extends AbstractAction
 {
     public function __construct()
     {
-        $this->name = "Отказаться";
-        $this->code = 5;
+        $this->name = "Начать задание";
+        $this->code = 1;
     }
 
     /**
      * Метод проверки прав
-     * Возвращает true в случае если текущий пользователь
-     * является заказчиком задания
+     * Возвращает true если текущий пользователь
+     * является исполнителем задания
      * @param int $executorId -- id исполнителя задания
      * @param int $customerId -- id заказчика задания
      * @param int $activeId -- id текущего пользователя
@@ -23,6 +23,6 @@ class ActionDeny extends AbstractAction
      */
     public function checkPermission(int $executorId, int $customerId, int $activeId): bool
     {
-        return $executorId !== $customerId && $customerId === $activeId;
+        return $executorId !== $customerId && $executorId === $activeId;
     }
 }
