+++ src/models/actions/ActionCancel.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace TaskForce\models;
+namespace TaskForce\models\actions;
 
-class ActionFinish extends AbstractAction
+class ActionCancel extends AbstractAction
 {
     public function __construct()
     {
-        $this->name = "Выполнено";
-        $this->code = 4;
+        $this->name = "Отменить";
+        $this->code = 2;
     }
 
     /**
      * Метод проверки прав
      * Возвращает true в случае если текущий пользователь
-     * является заказчиком задания
+     * является исполнителем задания
      * @param int $executorId -- id исполнителя задания
      * @param int $customerId -- id заказчика задания
      * @param int $activeId -- id текущего пользователя
@@ -23,6 +23,6 @@ class ActionFinish extends AbstractAction
      */
     public function checkPermission(int $executorId, int $customerId, int $activeId): bool
     {
-        return $executorId !== $customerId && $customerId === $activeId;
+        return $executorId !== $customerId && $executorId === $activeId;
     }
 }
